requirements.txt
BeautifulSoup4
ipython
git+https://github.com/django-nonrel/django@nonrel-1.5
git+https://github.com/django-nonrel/djangotoolbox
git+https://github.com/django-nonrel/mongodb-engine
djangorestframework
djangorestframework-gis
html2text
simplejson
ox
jsonfield
django-leaflet
django-geojson
django-cors-headers
python-geoip
drf_ujson
djangorestframework-csv
requests
python-dateutils
django-rest-swagger
